Rename IovLedgerWallet files
